Disable link-local addressing completely to address issue on radio port two
This is our current theory for the issues on CM3 introduced after we
switched to the 64-bit image. We will need someone to test. Another
option that I'd like to explore more in the future is using the static
IP fallback to make it easy for teams to debug networking issues; we
could probably fallback to something like 10.99.99.42.