Add swap to support pi3 series memory pressure
We observed fantastic pipeline performance on a Pi4.

Due to difficulty sourcing Pi4s, we did testing on a Pi3b+.

During testing, we found large lags when the pipeline stages that recognize a target run.  We also observed very slow GUI performance, and slow ssh connection.

At first, we assumed that CPU was the issue. However, using htop to observe what was going on, the CPUs were fine.  Memory pressure was the issue. The pi3b+ had exhausted all physical memory, and all 100M of the default swap.

We changed swap to 700MB, and after that change, the pi3b+ performs pipelines just as fast as the Pi4.  ( In fact a bit faster in our setup, but that's because the frame rate on our pi3 camera was faster.

I'm curious about why the set value of 100 was installed, though. I believe the original value ( dynamic) might have allowed the 3b+ to work fine.  Depending on why 100MB fixed was chosen, going back to dynamic might be the right answer
We didn't do any testing to find out if this value is the minimal size to solve the problem, so more or less memory might work. If it would be